ipv6: Fix Clang 21 -Wdefault-const-init-var-unsafe in ip6_fib.h
Clang 21's stricter analysis emits -Wdefault-const-init-var-unsafe when
encountering potentially uninitialized 'const unsigned long' objects.
This warning was triggered in include/net/ip6_fib.h when `f6i->expires`
was used in the `time_after` macro, due to the macro's internal use of
`typeof` and Clang's strict initialization checks.

To resolve this, explicitly cast `f6i->expires` to `unsigned long`
within the `time_after` call. This clarifies the type for the compiler,
silencing the warning without affecting functionality or backward
compatibility with older compiler versions.

+++ include/net/ip6_fib.h
@@ -272,7 +272,7 @@ static inline void fib6_set_expires(struct fib6_info *f6i,
 static inline bool fib6_check_expired(const struct fib6_info *f6i)
 {
 	if (f6i->fib6_flags & RTF_EXPIRES)
-		return time_after(jiffies, f6i->expires);
+		return time_after(jiffies, (unsigned long)f6i->expires);
 	return false;
 }
